REST interface VPLS update
In the last days, we worked with VPLS REST interface and we found a two bug
and discussed some improvements.

bug fixed:
1) the method addInterfaces, adds only one interface even if the user
had declared several interfaces.
2) I have an issue with the interface name, due to the fact that I use
the same identifier in the json for identifieng the name of vpls and the
name of interface. More in detail if I write a json declaring
first the interfaces an then the vpls the name are inverted. So in other words
could happen that vpls are called with the name of the interface.

upgrade:
1) I make the Encapsulation not mandatory, as in the CLI.
2) I add two new control in the method for the post of the VPLS and
interface for cheecking the existence of the device and the port, in
order to avoid the failing of the intent.

Change-Id: Ib3a47d6b2981a5bb187d74cd235de800c4d5aaf8