[ONOS-8123] Fix an issue related to the component properties.
Properties are not inherited by the derived classes. This means that
a subclass cannot change a property defined in the super class.

We are seeing an issue with FObj managers where the FlowObjectiveManager
is not enabled and ConfigAdmin suppresses the calls "cfg set" to its
properties. Moreover, the InOrderFlowObjective is not able to change
its properties because it does not receive calls to "modified".
The last issue happens because the properties defined in the Component
annotation are not inherited.

Additionally, this patch makes configurable the objective timeout in
InOrderFlowObjectiveManager.

Change-Id: Ibd84be914b15a17f82186b5013c0c0697d2657c3