[SDFAB-15] Allow delete with empty actions
Allow delete of flowrules with empty actions by using the stored flowentry attributes
to rebuild the original entry. This is possible as eveything in the flowrules store
is based on the flow id which is an hash of the match keys.

Improve the FabricUpfProgrammable by leveraging the improvements done
in the flowrule store. In particular, it removes the linear scan used
to find the original flowrule which is no longer necessary

Change-Id: I03a6efcdd4e70a7d55cb0757befd0f9b450ab718