[SDFAB-293] Fix UPF call setup performance degradation
It has been observed that calling setupBehavior and computeHardwareResources
for each flow introduce a not negligible delay and these operations were not
performed before the refactoring for each flow.

There is no need to call the setup steps of setupBehavior each time but should
be called only the very first time. This patch makes setupBehavior more smart.

Then, looking at the performance with the profiler, we noticed that computeHwResources
is the most expensive operation we have introduced recently - it does not seem necessary
to compute the hw resources each time. With this patch we compute the hw resources/limits
only at the behavior init.

Change-Id: If6352f25c23cf93c25fcf3cffae3d3834bc0e75f