[ONOS-8148][SDFAB-618] Avoid meter map distruction in cluster environment on instance restart
In general an instance restart should not lead to the deactivate method. Unfortunately,
the bundle refresh is changing the cards on the table. Please note that destroy()
behavior is implementation dependent.

In this patch we have fixed bundles that can be potentially affected by the
bundle refresh issue. However, other fixes may be needed in future and the fixes
are use cases dependent: it depends when the bundle refresh issue is triggered
along the chain of the deps.

Change-Id: I7a45439f74b255d8d4b1cbdb89b6dec8adf00687