[SDFAB-622] Protect phased recovery from wrong port config
ifacename is used as key when changing the admin state of
a port through GNMI. This patch guarantees that we use the
state contained in the device store to fill the gnmi blob.
This results to be necessary because the requests coming
from the north may be incomplete (without port name).
When this happens, PortNumber class fills the name with
the number which is wrong for the GNMI devices

Change-Id: I308538ae30cfe3fbf62d42c9a059114e8ff5b4b3