[SDFAB-887][SDFAB-893] Missig flowrules in the ECFlowRuleStore
Mastership changes can happen during the processing of the flow rules stats,
If there is no master while we are executing handleExistingFlow we will get
null when we try to retrieve the stored flow rule. This is wrongly handled
with the removal of the flow rule from the store which will cause lost
of connectivity. This patch basically removes the flow rule only when we are
sure of what we are doing, when we are not the code verifies if there are
proper conditions to continue the processing (check if there was a mastership
change).

Additionally, this patch improves the logging of the ECFlowRuleStore

Change-Id: I7b79e7ab3d8ccfa6edca6ba4ad5de93ada082265