Introduce allowlist for back to back duplicate events
If there are two duplicate back to back role requests
DeviceTaskExectur drops the duplicate response if they
are spaced by an interval < 1s. This patch introduces an
allowlist for event types that should not be dropped
by the back to back policy.

This patch fixes one scenario that led to SDFAB-1025

Change-Id: Ic62ba56228c9c0c2b368e8e68ff50d11e2dc4249